challenge_stefanini_python/requirements.txt
Django==2.0.3
djangorestframework==3.7.7
django-environ==0.4.5
